pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.6.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "sqlparse>=0.5.3",
]
